requirements.txt
imageio==2.9.0
opencv_contrib_python==4.2.0.32
dlib==19.21.0
Flask_Cors==3.0.8
imutils==0.5.3
requests==2.24.0
numpy==1.18.1
Flask==1.1.2
git+https://github.com/ageitgey/face_recognition_models
Pillow==8.0.1
